Add standalone Validator class
This class directly uses the in-memory representation of the LinkML schema to perform validation on rows of data. The data is represented as an array of arrays (which is Handsontable's native format). An array of column headers is also required at validation time in order to map list elements to slots.